Trace: Add a custom SysMin module for traces
Added module ti.trace.SysMin that replaces xdc.runtime.SysMin.
The new modules allows flexibility and replaces SysMin_putchar()
and SysMin_flush() functions. The flush() function is not used
since we implement the traces in a circular buffer. The circular
buffers warps around when it reaches the end of the buffer

Signed-off-by: Shahid Akhtar <sakhtar@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>