Configs: Use GPT3 and GPT4 as src clk for BIOS tick in Ducati
BIOS tick is currently sourced by the SysTick timer internal to
the M3 subsystem. This timer will run continuously only as long
as the Ducati functional clock, MPU_M3_CLK is running.

In order to save power, MPU_M3_CLK needs to be gated, and this
happens when the deep sleep mode configuration is set, and both
M3 cores execute WFI.

The SysTick timer stops counting when the MPU_M3_CLK is gated,
and result in all the BIOS timer-based functions like Task_sleep,
Semaphore_pend, Event_pend to never finish and wake up the
subsystem. This is avoided by choosing external gptimers as the
BIOS tick sources instead of the SysTick internal timer.

SysTick internal M3 timer is overridden with GPT3 in Core0.
SysTick internal M3 timer is overridden with GPT4 in Core1.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>