PM: Enable GPT3 and GPT4 interrupts as wakeup sources
Currently GPT3 and GPT4 are selected by the configuration files as
the source of the BIOS tick. This requires that these interrupts
also be configured as wakeup sources for the M3 subsytem. Changes
have been added to enable these interrupts in the WUGEN mask.

Note that configuring different gptimers as the source clock will
require similar changes to enable those interrupts in the WUGEN
mask.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>