Samples: Request GPT9 instead of GPT3 in resmgr sample
Since GPT3 and GPT4 are now used to provide BIOS tick, the
sample is returning an error when trying to request GPT3.
This has been changed to request another timer, GPT9, instead.

Watchdog may utilize the GPT9 in the future, and if so it
may need to be disabled for testing this.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>