Samples: Add a task to test HwSpinlock
A new file, hwspinlock_task.c is added and is used to test
various API provided by the HwSpinlock module like create &
delete; to lock & unlock.

Several tasks of different priorities are created in the test
task to compete against each other for different HwSpinlocks.
The HwSpinlock test task is added as part of test_omx test,
but is kept disabled by default.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>