Utils: Add the genextbin utility
A genextbin utility is added that processes the Ducati binary and
gives out an extended binary with the MMU L1 PTE table and the
firmware section TOC information needed by the signing tool. The
two generated sections are placed at specific physical address
locations and needs to be matched in the signing and authentication
tools.

Signed-off-by: Suman Anna <s-anna@ti.com>