Resources: Add 1D IO buffers to resource table and fix addresses
The 1D IO carveout buffers need to be mapped into the L2MMU, as
they won't be mapped into the TILER. This is added to the resource
table, so that it can be mapped into the MMU at the time of loading
and starting the ducati processors.

All the carveouts have been relocated to match the A9-side locations
for different sections (defined in the board files) to account for
the 1D IO carveout. The ducati image's code and ipc regions are now
assigned fixed physical addresses. The data section is immediately
before the 1D IO carveout area, and the IPC and code sections precede
the data section.

Signed-off-by: Suman Anna <s-anna@ti.com>