Platform: Rearrange the memory sections for DSP
Add specific sections for code, data, heap, Trace and PM
regions that mimic the arrangement made on Ducati. The
section sizes have also been increased to a default minimal
sizes.

This memory map is designed to use a total of 4MB of memory
- 1 MB for code & data, 2 MB for heap, and 1MB for all the
IPC sections.

The TRACE_BUF section is used to store the DSP traces, and
have them visible to the host processor. The PM_DATA section
is reserved specifically for code to save and restore the
context used to resume the DSP after hibernation.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>