Build: Generate a default firmware binary for DSP
Build for the C64T platform is enabled. The test_omx_dsp sample
is used as the default source to generate the tesla-dsp.bin image
for DSP.

For DSP, the entry point ti_sysbios_family_c64p_Hwi0 needs to be
passed explicitly as an argument to the image generator in order
to properly get the initial boot address.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>