IPC: Rearrange mboxId to procId mappings
MPU-side uses a pair of mailboxes for duplex communication between
MPU and a remote processor. The distribution is as follows:
    mailbox_1:
        mpu --> rproc1 : mbox 0
        mpu <-- rproc1 : mbox 1

    mailbox_2
        mpu <-- rproc2 : mbox 2
        mpu --> rproc2 : mbox 3

rproc1 is used for abstracting both M3 cores in one entity. mbox0
& mbox1 are properly reserved for Rx/Tx in Core0. However, mbox 2
is used for Rx in Core1 due to the rerouting logic. This is not
correct since mbox2 should be reserved for the second rproc, DSP.
This patch assigns mbox 4 for Rx in Core1 to avoid conflicts when
both remoteprocs (M3's and DSP) are enabled together.

The mbox nomenclature has also been modified to reflect the sending
direction.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>