Configs: Set Non-Posted policy for IPC & Peripheral regions
The IPC and Peripheral regions' policy has been marked to the
default NON_POSTED policy. This is done to ensure that all
r/w(s) within the IPC and peripheral registers reach the
destination and in the same order as intended/issued by the
CPU. This policy avoids the CPU getting any stray interrupts
from peripherals as the order is now guaranteed.

Signed-off-by: Suman Anna <s-anna@ti.com>