Samples: Update host sample to work with latest rpmsg-omx
Updated the omx_sample to be in line with the latest rpmsg-omx
implementation. The rpmsg-omx driver parses the data packet being
sent to perform verification and translation of any buffer handles.
The number of buffer handles being passed in the buffer is expected
at the beginning of the data payload in the omx packet.

Signed-off-by: Anagani Chandra <chandu@ti.com>
Signed-off-by: Subramaniam C.A <subramaniam.ca@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>