PM: Add wake lock/unlock APIs to control deep sleep mode
DeepSleep mode is used to gate the CPU clock when the M3 core
hits WFI. However, some internal resources like CTM and SysTick
timers are sourced from the CPU clock. So, when the CPU clk is
gated, these resources are also gated and stop working as soon
as the M3 Core idles.

The following API are provided to control the deep sleep mode.
Users must call these API in pairs.
    Void IpcPower_wakeLock();
    Void IpcPower_wakeUnlock();

The IpcPower_wakeLock() disables the deep sleep mode, and when
invoked, allows the internal resources to be functional as the
core is never gated. The IpcPower_wakeUnlock() restores the deep
sleep mode.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>