SrvMgr: Add support for invoking a disconnect hook function
The OMX_DISCONNECT message currently is completely processed within
the Service Manager module. There is no infrastructure support to
call a user-provided function so that the client can execute some
additional custom operations.

This patch adds the support for registering a user hook function with
the ServiceMgr module, which will be invoked upon a OMX_DISCONNECT
message. This hook function will be used to trigger resource cleanup
in case of abnormal or premature termination of a corresponding HLOS
process.

Signed-off-by: Abhishek Ranka <asranka@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>