DEH: Save raw exception registers into exception buffer
A subset of the SYS/BIOS Exception context data is saved in
raw binary form into the the exception buffer. This exception
buffer is accessible to the HLOS-side, and can be used as part
of a slave processor core dump. The exception buffer previously
had only a textual information regarding the exception type.

Signed-off-by: Shahid Akhtar <sakhtar@ti.com>