Resources: Split the IPC carveout memory
The IPC carveout region needs to have access permissions from both
A9 and Ducati. This region serves multiple purposes/functionalities
- the shared memory is used for IPC buffers, trace data, exception
or crash-dump data and Power Management data. This region needs to
be contiguous for defining a single firewall region.

The IPC carveout region currently is 2MB and is split into two 1MB
entries at different virtual addresses to address the limitations
of the M3 memory mapping attributes. The first entry will continue
to be used for the IPC VirtQueue buffers, while the second entry
will be used to contain the remaining IPC data buffers/memory. The
address for the second entry is chosen within a executable memory
region as per the M3 memory map attributes, so that the power
management data can be moved into this entry region as well.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>