Trace: Configure timestamp format based on BIOS tick period
The timestamp currently being printed in the traces is programmed
to work with only a BIOS tick period of 1ms. The BIOS tick period
is a configurable option, and if changed, the current trace
timestamps would be incorrect.

Changes have been made to compute the timestamp based on the value
of the tick period. The current changes would support a range of
tick period values from 1 us to 1 sec. Anything beyond would need
additional changes, but 1 sec is already a decently large tick
period value for various Clock related functions.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>