Trace: Add a new StackDbg module
A new module StackDbg provides API to unwind the stack natively, without
the aid of symbolic debugging information and without the usage of a frame
pointer.  The algorithm used is heuristic, and so the stack trace may be
partial or incorrect.  For the most past, though, the algorithm produces an
accurate trace, and when it does not, it is obvious that the stack trace is
wrong.

The stack-walking function takes as an argument a visit function, which is
called for every potential stack entry.  The visit function may choose to print
out that entry, or to add it to a stack-trace buffer for later printing, or to
ignore it.  The visit function also tells the stack-trace function for each
entry, whether the stack-walking should continue or stop.  This way a caller of
the stack-walking function may unwind the stack only as much as it necessary.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Shahid Akhtar <sakhtar@ti.com>
Signed-off-by: Iliyan Malchev <malchev@google.com>