Utils: Append version in firmware header
Use git describe command to identify the core0, core1 and
sysbios-rpmsg version used to create the firmware binary.

The M3 baseimages (xem3's) are expected to reside in a git
repository. Copying the same binaries to an external folder
and generating the firmware binary using the baseimages from
that external folder will not give out the desired information.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>