IPC: Add a function to notify host upon init completion
In some situations, it is good to let the host know when the
boot has completed a certain basic initialiazation.

Add a function VirtQueue_postInitDone() and the message
RP_MSG_BOOTINIT_DONE to let the host know that the processor
core has finished the initialization during the boot process.
This can be used by the host processor to use some constraints
to speed up the loading and booting of the slave processors.

Signed-off-by: Miguel Vadillo <vadillo@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>