SrvMgr: Send boot init completion notification to host
Considering the creation of the rpmsg-omx channels as the point
when the booting/initialization is complete, send a special
notification message to the host.

This message is sent from just the AppM3 (in case of IPU) and DSP
cores, so that the host side can get notified by the core that
completes the initialization last.

Signed-off-by: Miguel Vadillo <vadillo@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>