Build: Update to newer XDC and IPC toolset
The build system is updated to use newer XDC (XDC 3.23.02.47) and
IPC (IPC 1.24.02.27) tools.

This newer toolset are the latest on the respective release streams,
and allow a single unified version of XDC and IPC tools between
the SMP and non-SMP versions.

The README has also been updated with the corresponding download
links.

Signed-off-by: Suman Anna <s-anna@ti.com>