Utils: Assign an entry point for DSP when not provided
Symbol address of ti_sysbios_family_c64p_Hwi0 is assigned
as the entry point for DSP in genrprc utility, when no entry
point information is passed to genrprc utility. This symbol
branches off to the c_int00, and also satifies the 1K alignment
requirement for the DSP boot address.

This change is done in genrprc so that base images built outside
of sysbios-rpmsg can easily be used to generate a tesla binary
without having to specify any special entry point, in case if
they are not aware of the alignment restrictions. An user passed
entry point is still honored to retain the flexibility.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>