ResMgr: Use NameMap for connections and disconnections
The ResMgr channel is not created by default on the host-side
using a fixed channel anymore. Use the NameMap framework to
create the service end-point on the host-side.

NOTE: The NameMap framework is intended for publishing the
services that the slave processor provides, and ResMgr is
really a server (or a service providing resource management)
on the host-side. There is currently no framework for publishing
services from the host-side to the remote-side. The above
usage of NameMap allows the resource manager driver to be
created dynamically, and this needs to be revisited later
in the context of service publishing from any processor to
another.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>