Build: Improve flexibility for picking tool versions
The Makefile has been enhanced to be able to pick the BIOS
tool versions through parameters to the 'make' command. The
previous method of choosing custom versions through exported
shell environment variables is still valid. The parameters
to 'make' are the same as the environment variables, and
would override the corresponding environment variables.

Signed-off-by: Suman Anna <s-anna@ti.com>