Samples: Update host samples to work for both Ducati & DSP
Updated both the omx_sample and omx_benchmark sample source code to
be able to take in an additional argument to run the sample with
either of the Ducati processors or the DSP processor.

The scripts and README have also been updated. Changes are also made
to have these files copied one level down from where they are currently
required to be copied.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Chandra Sekhar.Anagani <chandu@ti.com>