Configs: Move Idle module from IpcCommon to specific cfg files
Idle module interface is different for SMP platform. SMP requires
that Idle functions be added to each of the cores, and if it is the
same Idle function, then it should be a thread-safe function. The
inclusion of the module is therefore moved from the generic
IpcCommon.cfg.xs file to each one of the specific core configuration
files.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Paul Hunt <hunt@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>