Resources: Add SMP-mode support to resource table
Add a 'SMP' macro based on the platform information to support
SMP-mode on the IPU sub-system. This will list just one trace
buffer, one crash dump buffer and just one entry point as the
two IPU processors are treated as a single processor system
in SMP mode.

Libraries for SMP and non-SMP mode are built simultaneously
during build time, and the appropriate library is picked up
during the link stage in the final executable.

Signed-off-by: Paul Hunt <hunt@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>