IPC: Add dependency towards pm package in rpmsg package
When Idle functions are taken away or commented out from the common
configuration, the compiler is throwing a linker error (not finding
PM symbols) as it does not see any explicit dependency due to linker
optimizations.

An explicit dependency is added in the rpmsg package towards the pm
package to alleviate this problem.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>