SrvMgr: Use rpmsg-omx1 channel name for IPU in SMP mode
The IPU subsystem will create only one rpmsg channel in SMP
mode. The name 'rpmsg-omx1' is used to publish the channel
name in SMP mode. This helps to keep backward compatibility
with the corresponding driver and userspace applications.

Simultaneous support is provided for both non-SMP and SMP
modes through a compiler preprocessor macro. Libraries for
both the modes are built simultaneously during build time,
and the appropriate library is picked up during the linking
stage in the final executable based on the platform name.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>