Trace: Customize tracing for SMP-mode
The tracing code has been enhanced to support the SMP-mode.
In SMP-mode, each processor core will have its own line
buffer and the traces are output into this linebuffer until
a new line is output or the line buffer has been filled up.
The respective line buffer is flushed out into the shared
output buffer with the processor core id prepended to the
respective line buffer trace.

Libraries for SMP and non-SMP mode are built simultaneously
during build time, and the appropriate library is picked up
during the linking stage in the final executable based on
the platform name.

Signed-off-by: Paul Hunt <hunt@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>