ResMgr: Change request and ack structures
The Resource requests are changed from requesting by id to requesting
by name to match the latest kernel-side code. The translation from the
resource id to the name is done within the IpcResource module, by
maintaining the matching name list. This change is transparent, and
there is no change in the exported API interface.

Also, the request and ack structures are split into sub-structures
depending on the type of request and the ack response for a specific
request. That is because each action needs its own parameters.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>