ResMgr: Add local structures for auxclk and regulator resources
The auxclk resource on the host-side now receives the name of the
auxclk being requested as a parameter (to avoid the limit on num
of auxclk that can be requested and to avoid hardcoding that in
the host side). Therefore, now the auxclk structure is different
and we need to create a local auxclk structure match the host side
and to keep the original structure to match exported APIs.

The same applies to the regulator resource.

The old structures can be removed once all users of IpcResource are
migrated to the new changes.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>