PM: Configure deepsleep for IPU in a generic way
Depending on the BIOS version used, the Power Suspend API for IPU
could be executed as an asynchronous non-blocking call in just one
core (in a SMP environment) or as a synchronous blocking call in
both cores (non-SMP environment). This patch ensures that each
core sets/recovers its deepsleep configuration irrespective of the
mode and the core where the Power_suspend function is actually
executed.

Signed-off-by: Paul Hunt <hunt@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>