PM: Add SMP support to the Power Management code
The SMP support is added to the core PM module for IPU,
IpcPower. The Power Management functionality is adjusted
for the lack of two independent processor cores in the
IPU. The simultaneous support for non-SMP and SMP modes
are provided through a compiler preprocessor macro.

Libraries for SMP and non-SMP mode are built simultaneously
during build time, and the appropriate library is picked up
during the linking stage in the final executable based on
the platform name.

Signed-off-by: Paul Hunt <hunt@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>