Build: Add build support for IPU platforms in SMP-mode
The build for SMP-mode IPU target is enabled for both OMAP4
and OMAP5 platforms by configuring the build system to an
exclusive SMP build mode:
   bash$ make smp_config

The proper BIOS and XDC versions that support SMP-mode are
selected automatically when the above configuration is chosen
overriding the non-SMP versions. Only the IPU SMP targets are
built upon issuing the usual build command:
   bash$ make

For going back to the default non-smp configuration
next command should be used:
   bash$ make unconfig

The SMP and regular mode binaries can be built by invoking
the build twice without doing a clean build, but only
changing the config.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Paul Hunt <hunt@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>