IPC: Fix a race condition in MessageQCopy initialization
When a task called MessageQCopy_init function, it immediately
set the curInit variable to true before starting the actual
initialization. If another task called the same function, the
curInit was true already, and proceeded as if the initialization
was already done, while the module is still being initialized.

A semaphore initialized with its count to 1 is used as a mutual
exclusion protection between the competing tasks.

Signed-off-by: Paul Hunt <hunt@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>