PM: Add new API to control suspend process
Added new API to allow applications/tasks on IPU or DSP to
control the suspend ability on the processors they are
executing on. The slave processors ignore the SUSPEND message
if user has disabled suspend.

These API compliment the existing IpcPower_wakeLock and
IpcPower_wakeUnlock API, but also allow the clocks to be cut
for better power savings.

New API added are:
UInt IpcPower_hibernateLock()
UInt IpcPower_hibernateUnlock()
Bool IpcPower_canHibernate()

The first 2 API return the current number of tasks holding
the suspend lock, and the last API returns if the processor
is allowed to suspend or not.

Signed-off-by: Shahid Akhtar <sakhtar@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>