Samples: Update the ResMgr sample to check for proper status
Resmgr Sample is modified to check for proper return status for
each of the resource request and release function calls. The sample
now prints an easy-to-read succeeded or failed status with the
corresponding status codes for various combinations of the resource
and constraints requests.

The sample has also been updated properly to support the different
resource request variations for OMAP5.

Signed-off-by: Chandra Sekhar.Anagani <chandu@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>