IPC: Add SMP support to the VirtQueue module
The SMP support in IPU is added to the core IPC transport
layer - the VirtQueue module. For SMP, we only need one
pair of VirtQueues as if they were only one core. Support
for non-SMP and SMP is simultaneously provided by using a
compiler preprocessor macro.

Libraries for SMP and non-SMP mode are built simultaneously
during build time, and the appropriate library is picked up
during the link stage in the final executable.

Signed-off-by: Paul Hunt <hunt@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>