IPC: Simplify VirtQueue init sequence
Simplify the VirtQueue init sequence by removing the init
stage code and the semaphore init sequence code.

The starting device address for the vrings and associated
buffers is hard-coded and is expected to be the same for
all remote processors. It is perfectly fine to hard code
device address, since it's our address space, and this is
actually what we've been doing.

Signed-off-by: Ohad Ben-Cohen <ohad@wizery.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>