IPC: Correct VirtQueue_getAvailBuf based on review
A review of VirtQueue revealed the index to the available ring
buffer was being calculated differently than lguest from which
it was modelled. It worked given the current usage model, but
to be more correct, it should be updated to match what Linux
side does.

This change was verified to work by asserting that head index
was the same in both cases under sample tests. A minor trace
change to display the ring.avail addr is also added.

Signed-off-by: G Anthony <a0783926@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>