Trace: Introduce line buffer for local caching
Introduce a local line buffer to which the characters are output.
The line buffer is flushed to the external shared buffer only
when the line is filled up or when a new line is encountered.

This change re-introduces back the actual flush functionality
(without a new line) present in the baseline SysMin module.
It also allows the future scalability for a SMP system wherein
the individual cores can write to their own line buffers without
mangling the actual output buffer.

Signed-off-by: Paul Hunt <hunt@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>