ResMgr: Build Resource Manager for both SMP and non-SMP modes
Two different ResMgr libraries are built for both non-SMP and SMP
modes at build time based on configuration, and the appropriate
library is picked up during the linking stage in the final
executable based on the platform name.

The conditional build allows the build configuration to change
without overwriting the other mode's library.

Signed-off-by: Suman Anna <s-anna@ti.com>