DEH: Move out Watchdog functionality into a separate module
The Watchdog functionality has been separated out from the DEH
module and moved into a separate Watchdog module. The Deh files
will be used only for the exception reporting code that is
specific to each target processor. The new Watchdog module is
currently supported only for OMAP4/OMAP5 devices, but is designed
for scalability.

The patch also includes generic cleanup in terms of code format
and comments.

Signed-off-by: Shahid Akhtar <sakhtar@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>