Trace: Use proper locking for trace buffers in SMP mode
SysMin has a line buffer for each core and a module-wide
shared buffer, common to both the cores and used to flush
out line buffers one at a time. The tracing functionality
depends on locking of these buffers properly. The line buffer
is protected by disabling local core interrupts, and the
shared buffer is protected by acquiring a system lock.

The current code is re-enabling the local core interrupts
before checking whether the line buffer needs flushing. This
opens up a small window for the line buffer to be overwritten
by a different task or interrupt while the current thread is
waiting for the system lock to flush the core line buffer.
This is fixed by enabling the local core interrupts properly
after checking on the flush condition. Releasing the system
gate re-enables the core interrupts automatically, so there
is no need for enabling local interrupts specifically when
flushing the line buffer.

Signed-off-by: Suman Anna <s-anna@ti.com>