DEH: Enhance Watchdog module to support SMP-mode
The Watchdog module has been modified to support the multicore
SMP-mode. The Watchdog timers for all cores are initalized at
the same time. The WDT for Core0 is same as before. For Core1,
the WDT is stopped during Idle task and restarted when a task
is scheduled on Core1.

The code uses ti.sysbios.hal.Core from SMP/BIOS to determine
which core the code is running on in SMP-mode.

Signed-off-by: Shahid Akhtar <sakhtar@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>