PM: Enable mbox interrupt in wugen for DSP before suspending
The DSP is not resuming properly for some reason with the latest
host-side code. The hard-reset is being effective only after the
DSP is specifically woken up. This is currently being achieved by
enabling the mailbox interrupt in the DSP Wakeup Generator (wugen)
before suspending. A mailbox message always triggers the resume on
the host processor side, and this message will wake up the DSP and
cause the resume to complete successfully.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>